+++ Sources/Requests/Requests+PingReq.swift
@@ -1,0 +1,34 @@
+import Foundation
+
+extension Requests {
+    public struct PingReq: Codable, Hashable, Sendable {
+        public let nonce: String?
+        /// Additional properties that are not explicitly defined in the schema
+        public let additionalProperties: [String: JSONValue]
+
+        public init(
+            nonce: String? = nil,
+            additionalProperties: [String: JSONValue] = .init()
+        ) {
+            self.nonce = nonce
+            self.additionalProperties = additionalProperties
+        }
+
+        public init(from decoder: Decoder) throws {
+            let container = try decoder.container(keyedBy: CodingKeys.self)
+            self.nonce = try container.decodeIfPresent(String.self, forKey: .nonce)
+            self.additionalProperties = try decoder.decodeAdditionalProperties(using: CodingKeys.self)
+        }
+
+        public func encode(to encoder: Encoder) throws -> Void {
+            var container = encoder.container(keyedBy: CodingKeys.self)
+            try encoder.encodeAdditionalProperties(self.additionalProperties)
+            try container.encodeIfPresent(self.nonce, forKey: .nonce)
+        }
+
+        /// Keys for encoding/decoding struct properties.
+        enum CodingKeys: String, CodingKey, CaseIterable {
+            case nonce
+        }
+    }
+}