+++ Tests/Wire/Resources/Shipping/TaskGroups/Expanded/TaskGroupsExpandedClientWireTests.swift
@@ -1448,6 +1448,9 @@ import Chrt
             page: 1,
             pageSize: 1,
             search: "search",
+            filterStatus: [
+                .draft
+            ],
             filterDraftStartedAtTimestampLte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
             filterDraftStartedAtTimestampGte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
             filterStagedAtTimestampLte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
@@ -1586,6 +1589,9 @@ import Chrt
             page: 1,
             pageSize: 1,
             search: "search",
+            filterStatus: [
+                .draft
+            ],
             filterDraftStartedAtTimestampLte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
             filterDraftStartedAtTimestampGte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
             filterStagedAtTimestampLte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),